requirements.txt
motor
databases
authlib
jwt
pyjwt
google-auth
asyncio
httpx
gunicorn
python-multipart
pandas
pyexcel-xlsx
pyexcel
fastapi[all]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
loguru
sqlalchemy
asyncpg
sqlmodel
alembic
psycopg2-binary
requests
stripe